Workaround for `JavaScript::Packer` Bug
The `best` setting chooses the better of `shrink` and `obfuscate`, but
it has to do twice the work for this. Also, `obfuscate` does not
actually work with our JavaScript, as empirically determined.

Unfortunately for us, `best` has so far been choosing `shrink` every
time, so we did not notice it. It took a manual prodding to reveal all
that was wrong with this setup.