Force use of Brotli (`Content-Encoding: br`)
This will _most likely_ not make it anywhere near `master` without some
additional commits on top, not because the author has any sympathies for
browsers that don’t support Brotli, but because there are debugging
tools which can’t and mostly don’t need to understand Brotli, and
(relatively) dumb clients that just need to access the JSON API without
bothering about decoding it.

Then again, all of the above might just seem silly when (if) this does
end up naked on `master`.